rna parse: traverse all occurences of the MID_SEQ
Because the 9bp MID_SEQ is shorther than the 11bp random promoter
region, it is likely that the first occurence of the MID_SEQ is acutally
the randome promoter region. This will cause false positive and false
negative.

False positive: END_SEQ can be found in the downstream due to the
randomness of 17bp barcode. Increased the length of END_SEQ to reduce
the rate of this situation.

False negative: The real MID_SEQ is at the downstream. Look for all
occurences of MID_SEQ to reduce the rate of this situation.